Reduce ffmpeg size (replace full-gpl version with audio)

+++ DownloaderAppMobile/DownloaderAppMobile.Android/Services/FFmpeg.cs
@@ -35,7 +35,6 @@ namespace DownloaderAppMobile.Droid
             using var session = FFmpegSession.Create(args);
             session.AddLog(new Log(session.SessionId, Level.AvLogVerbose, string.Empty));
 
-
             await Task.Run
             (
                 () =>
@@ -57,7 +56,7 @@ namespace DownloaderAppMobile.Droid
             try
             {
                 // create valid front cover for audio
-                File.WriteAllBytes(tempCoverFilePath, options.Thumbnail);
+                await File.WriteAllBytesAsync(tempCoverFilePath, options.Thumbnail);
 
                 string command = string.Format(VIDEO_TO_AUDIO_METADATA_FORMAT, tempVideoFilePath, tempCoverFilePath,
                     string.Format(DEFAULT_METADATA_FORMAT, new object[]
@@ -69,7 +68,11 @@ namespace DownloaderAppMobile.Droid
                         options.AlbumArtist
                     }), destinationFilePath);
 
-                File.Create(destinationFilePath).Dispose();
+                if (!File.Exists(destinationFilePath))
+                {
+                    File.Create(destinationFilePath).Dispose();
+                }
+
                 await ExecuteAsync(command, cancellationToken);
             }
             finally
@@ -97,7 +100,11 @@ namespace DownloaderAppMobile.Droid
                         options.AlbumArtist
                     }), destinationFilePath);
 
-                File.Create(destinationFilePath).Dispose();
+                if (!File.Exists(destinationFilePath))
+                {
+                    File.Create(destinationFilePath).Dispose();
+                }
+
                 await ExecuteAsync(command, cancellationToken);
             }
             finally
